OrientDB DAO (deprecated)

+++ src/main/java/pers/yozora7/lanfirewallmgr/parser/data/Rule.java
@@ -13,7 +13,7 @@ import java.util.Set;
 @AllArgsConstructor
 public class Rule {
     // 序号
-    private int id;
+    private long id;
     // 名称
     private String name;
     // 源端安全域
@@ -21,15 +21,15 @@ public class Rule {
     // 源端地址集列表 (source-address address-set ...)
     private Set<String> srcSets;
     // 源端IP地址列表
-    private Set<Integer> srcAddressIds;
+    private Set<Long> srcAddressIds;
     // 目标端安全域
     private String dstZone;
     // 目标端地址集列表
     private Set<String> dstSets;
     // 目标端IP地址列表
-    private Set<Integer> dstAddressIds;
+    private Set<Long> dstAddressIds;
     // 服务列表
-    private Set<Integer> serviceIds;
+    private Set<Long> serviceIds;
     // service-group
     private Set<String> serviceGroups;
     // action
